+++ st2reactor/tests/unit/test_enforce.py
@@ -15,6 +15,11 @@
 
 from __future__ import absolute_import
 
+# pytest: make sure monkey_patching happens before importing mongoengine
+from st2common.util.monkey_patch import monkey_patch
+
+monkey_patch()
+
 import mock
 
 from st2common.constants import action as action_constants